requirements.txt
# requirements.txt
torch>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pinterest-api-python>=0.1.0
python-dotenv>=0.19.0
google-generativeai>=0.1.0
streamlit>=1.10.0